requirements.txt
Flask==2.3.3
tensorflow==2.16.1
keras==3.3.3
numpy==1.24.3
Pillow==10.1.0
werkzeug==2.3.7